safari-600.7-branch/LayoutTests/editing/text-iterator/count-matches-in-form.html

<head>
<script src="../../resources/js-test-pre.js"></script>
<body>
<fieldset><input value="rule"></fieldset>
<script>
shouldBe("internals.countMatchesForText('rule', 23, '')", "1");
</script>
<script src="../../resources/js-test-post.js"></script>
