safari-600.7-branch/LayoutTests/editing/text-iterator/count-matches-in-form-expected.txt

PASS internals.countMatchesForText('rule', 23, '') is 1
PASS successfullyParsed is true

TEST COMPLETE
